requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
appnope==0.1.0
astroid==2.4.0
async-timeout==3.0.1
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
Brotli==1.0.7
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
cfscrape==2.1.1
chardet==3.0.4
click==7.1.2
cloudscraper==1.2.46
colorama==0.4.3
console-menu==0.6.0
cssselect==1.1.0
CursesMenu==0.0.1a0
decorator==4.4.2
dhooks==1.1.3
discord==1.0.1
discord.py==1.3.3
fake-headers==1.0.2
fake-useragent==0.1.11
Faker==4.0.3
Flask==1.1.2
html5lib==1.0.1
idna==2.9
ipykernel==5.3.4
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jupyter-client==6.1.6
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.5
names==0.3.0
parse==1.15.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
polling==0.3.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
pyee==7.0.2
Pygments==2.6.1
pylint==2.5.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
python-cfonts==1.3.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
requests==2.23.0
requests-html==0.10.0
requests-toolbelt==0.9.1
six==1.14.0
soupsieve==2.0.1
spotipy==2.13.0
text-unidecode==1.3
toml==0.10.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
urllib3==1.25.9
w3lib==1.22.0
wcwidth==0.2.4
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2